fixed image display, extracted PrizeCard, optimised styles
